Refactor(views): Moves route enums to views
Moves enums that represent routes to the views that use them.
This makes the code more modular and easier to maintain.
