--- sonder/Routes/CircleFeedItemRoute.swift
@@ -1,12 +1,0 @@
-//
-//  CircleFeedItemRoute.swift
-//  sonder
-//
-//  Created by Bryan Sample on 12/15/25.
-//
-
-import SonderDTOs
-
-enum CircleFeedItemRoute {
-    case post, event
-}
