--- sonder/Routes/CircleOnboardingRoute.swift
@@ -1,12 +1,0 @@
-//
-//  CircleOnboardingRoute.swift
-//  sonder
-//
-//  Created by Bryan Sample on 12/3/25.
-//
-
-import SonderDTOs
-
-enum CircleOnboardingRoute {
-    case create, invite
-}
