--- sonder/Routes/LandingPageRoute.swift
@@ -1,10 +1,0 @@
-//
-//  LandingPageRoute.swift
-//  sonder
-//
-//  Created by Bryan Sample on 12/5/25.
-//
-
-enum LandingPageRoute {
-    case feed, calendar, post, circles, profile
-}
